requirements.txt
torch==2.1.2
Pillow==10.1.0
streamlit==1.29.0
numpy==1.26.2
opencv_python==4.8.0.76
opencv_python_headless==4.8.1.78
```